% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors_daily.R
\name{aqr_bab_daily}
\alias{aqr_bab_daily}
\title{Get the Bet Against Beta Factor}
\usage{
aqr_bab_daily(.tidy = TRUE)
}
\arguments{
\item{.tidy}{A flag. Should the output be tidy? The default is \code{TRUE}.}
}
\value{
A \code{tibble}.
}
\description{
Downloads data with the excess returns of long/short equity Betting Against Beta
(BAB) factors.
}
\examples{
if (FALSE) {
  aqr_bab_daily()
}
}
\references{
Frazzini, Andrea and Lasse H. Pedersen (2014),
'Betting Against Beta', Journal of Financial Economics, 111.
}
