% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commodities.R
\name{aqr_commodities_long_run}
\alias{aqr_commodities_long_run}
\title{Long Run Commodities Performance}
\usage{
aqr_commodities_long_run(.tidy = TRUE)
}
\arguments{
\item{.tidy}{A flag. Should the output be tidy? The default is \code{TRUE}.}
}
\value{
A \code{tibble}.
}
\description{
Downloads an updated version of the data used in “Commodities for the Long Run".
}
\examples{
if (FALSE) {
  aqr_commodities_long_run()
}
}
\references{
“Commodities for the Long Run" (Levine, Ooi, Richardson,
and Sasseville, FAJ, 2018)
}
