% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit_risk.R
\name{aqr_credit_risk_premium}
\alias{aqr_credit_risk_premium}
\title{Credit Risk Premium}
\usage{
aqr_credit_risk_premium(.tidy = TRUE)
}
\arguments{
\item{.tidy}{A flag. Should the output be tidy? The default is \code{TRUE}.}
}
\value{
A \code{tibble}.
}
\description{
Downloads the credit excess return data used in Asvanunt and Richardson, 2015,
“The Credit Risk Premium," working paper.
}
\examples{
if (FALSE) {
  aqr_credit_risk_premium()
}
}
