% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/premia.R
\name{aqr_factor_premia_monthly}
\alias{aqr_factor_premia_monthly}
\title{Premia From a Wide Array of Factors}
\usage{
aqr_factor_premia_monthly(.tidy = TRUE)
}
\arguments{
\item{.tidy}{A flag. Should the output be tidy? The default is \code{TRUE}.}
}
\value{
A \code{tibble}.
}
\description{
Downloads data with the  self-financing returns of equity, bonds, currencies and
commodities.
}
\examples{
if (FALSE) {
  aqr_factor_premia_monthly()
}
}
\references{
Ilmanen, Israel, Lee, Moskowitz & Thapar (2021),
“How Do Factor Premia Vary Over Time? A Century of Evidence".
}
