% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors_monthly.R
\name{aqr_hml_devil_monthly}
\alias{aqr_hml_devil_monthly}
\title{Get the High Minus Low Factor}
\usage{
aqr_hml_devil_monthly(.tidy = TRUE)
}
\arguments{
\item{.tidy}{A flag. Should the output be tidy? The default is \code{TRUE}.}
}
\value{
A \code{tibble}.
}
\description{
Downloads data with the  self-financing returns of equity High Minus Low Devil
(HML Devil) factors using the Book to Market definition of Asness and Frazzini (2013).
}
\examples{
if (FALSE) {
  aqr_hml_devil_monthly()
}
}
