% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors_daily.R
\name{aqr_hml_ff_daily}
\alias{aqr_hml_ff_daily}
\title{Get the High Minus Low Factor}
\usage{
aqr_hml_ff_daily(.tidy = TRUE)
}
\arguments{
\item{.tidy}{A flag. Should the output be tidy? The default is \code{TRUE}.}
}
\value{
A \code{tibble}.
}
\description{
Downloads data with the self-financing returns of equity High Minus Low (HML)
factors using the Book to Market definition of Fama and French (1992).
}
\examples{
if (FALSE) {
  aqr_hml_ff_daily()
}
}
