% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors_daily.R
\name{aqr_mkt_daily}
\alias{aqr_mkt_daily}
\title{Get the Market Factor}
\usage{
aqr_mkt_daily(.tidy = TRUE)
}
\arguments{
\item{.tidy}{A flag. Should the output be tidy? The default is \code{TRUE}.}
}
\value{
A \code{tibble}.
}
\description{
Downloads data with the market returns in excess of t-bills.
}
\examples{
if (FALSE) {
  aqr_mkt_daily()
}
}
