% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum.R
\name{aqr_momentum_monthly}
\alias{aqr_momentum_monthly}
\title{Get the AQR Momentum Indexes}
\usage{
aqr_momentum_monthly(.tidy = TRUE)
}
\arguments{
\item{.tidy}{A flag. Should the output be tidy? The default is \code{TRUE}.}
}
\value{
A \code{tibble}.
}
\description{
Downloads data with the returns of AQR Momentum Index, AQR Small Cap Momentum Index,
and AQR International Momentum Index.
}
\examples{
if (FALSE) {
  aqr_momentum_monthly()
}
}
