% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors_daily.R
\name{aqr_qmj_daily}
\alias{aqr_qmj_daily}
\title{Get the Up Quality Minus Junk Factor}
\usage{
aqr_qmj_daily(.tidy = TRUE)
}
\arguments{
\item{.tidy}{A flag. Should the output be tidy? The default is \code{TRUE}.}
}
\value{
A \code{tibble}.
}
\description{
Downloads data with the self-financing returns of of a long/short Quality
Minus Junk (QMJ) factors.
}
\examples{
if (FALSE) {
  aqr_qmj_daily()
}
}
