% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors_monthly.R
\name{aqr_qmj_monthly}
\alias{aqr_qmj_monthly}
\title{Get the Quality Minus Junk Factor}
\usage{
aqr_qmj_monthly(.tidy = TRUE)
}
\arguments{
\item{.tidy}{A flag. Should the output be tidy? The default is \code{TRUE}.}
}
\value{
A \code{tibble}.
}
\description{
Downloads data with the self-financing returns of a of a long/short Quality
Minus Junk (QMJ) factors.
}
\examples{
if (FALSE) {
  aqr_qmj_monthly()
}
}
