% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors_monthly.R
\name{aqr_smb_monthly}
\alias{aqr_smb_monthly}
\title{Get the Small Minus Big Factor}
\usage{
aqr_smb_monthly(.tidy = TRUE)
}
\arguments{
\item{.tidy}{A flag. Should the output be tidy? The default is \code{TRUE}.}
}
\value{
A \code{tibble}.
}
\description{
Downloads data with the self-financing returns of equity Small Minus Big (SMB)
factors.
}
\examples{
if (FALSE) {
  aqr_smb_monthly()
}
}
