% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors_monthly.R
\name{aqr_umd_monthly}
\alias{aqr_umd_monthly}
\title{Get the Up Minus Down Factor}
\usage{
aqr_umd_monthly(.tidy = TRUE)
}
\arguments{
\item{.tidy}{A flag. Should the output be tidy? The default is \code{TRUE}.}
}
\value{
A \code{tibble}.
}
\description{
Downloads data with the self-financing returns of equity Up Minus Down (UMD) factors.
}
\examples{
if (FALSE) {
  aqr_umd_monthly()
}
}
